+++ src/neo4j_genai/retrievers/__init__.py
@@ -12,3 +12,32 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+
+from .hybrid import HybridCypherRetriever, HybridRetriever
+from .text2cypher import Text2CypherRetriever
+from .vector import VectorCypherRetriever, VectorRetriever
+
+
+__all__ = [
+    "VectorRetriever",
+    "VectorCypherRetriever",
+    "HybridRetriever",
+    "HybridCypherRetriever",
+    "Text2CypherRetriever",
+]
+
+
+try:
+    from .external.pinecone.pinecone import PineconeNeo4jRetriever  # noqa: F401
+
+    __all__.append("PineconeNeo4jRetriever")
+except ImportError:
+    pass
+
+
+try:
+    from .external.weaviate.weaviate import WeaviateNeo4jRetriever  # noqa: F401
+
+    __all__.append("WeaviateNeo4jRetriever")
+except ImportError:
+    pass